# paymentservice/src/main/resources/openapi/payment-service-openapi.yaml
openapi: 3.0.1
info:
  title: Payment Service API
  description: |
    Wallet and payment management service with Stripe integration.
    Handles top-up, debit, payment processing, refunds, and invoice generation.
  version: 1.0.0
  contact:
    name: HomeGenie Team
servers:
  - url: http://localhost:8083
    description: Local development
  - url: http://production-url:8083
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /api/payments/create:
    post:
      tags:
        - Payment Operations
      summary: Create payment
      description: Create payment for maintenance request
      operationId: createPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCreateRequest"
      responses:
        "200":
          description: Payment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /api/payments/{paymentId}:
    get:
      tags:
        - Payment Operations
      summary: Get payment by ID
      operationId: getPaymentById
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Payment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
...
  /api/payments/invoices/create:
    post:
      tags:
        - Invoice Operations
      summary: Create invoice
      description: Generate invoice for payment
      operationId: createInvoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceCreateRequest"
      responses:
        "200":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
  /api/payments/invoices/{invoiceId}:
    get:
      tags:
        - Invoice Operations
      summary: Get invoice by ID
      operationId: getInvoiceById
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
  /api/payments/invoices/payment/{paymentId}:
    get:
      tags:
        - Invoice Operations
      summary: Get invoice by payment ID
      operationId: getInvoiceByPaymentId
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"

  # Health Check (1 endpoint)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaymentCreateRequest:
      type: object
      required:
        - userId
        - maintenanceRequestId
        - amount
      properties:
        userId:
          type: integer
          format: int64
          example: 1
        maintenanceRequestId:
          type: integer
          format: int64
          example: 1
        amount:
          type: number
          format: double
          example: 500000.00
        description:
          type: string
          example: "Sửa điều hòa"
        paymentMethod:
          type: string
          enum: [WALLET, STRIPE, CASH]
          example: "WALLET"
    PaymentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        userId:
          type: integer
          format: int64
          example: 1
        maintenanceRequestId:
          type: integer
          format: int64
          example: 1
        amount:
          type: number
          format: double
          example: 500000.00
        status:
          type: string
          enum: [PENDING, PROCESSING, COMPLETED, FAILED, REFUNDED]
          example: "COMPLETED"
        paymentMethod:
          type: string
          enum: [WALLET, STRIPE, CASH]
          example: "WALLET"
        transactionId:
          type: string
          example: "txn_123456789"
        description:
          type: string
          example: "Sửa điều hòa"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"
        paidAt:
          type: string
          format: date-time
          nullable: true
          example: "2024-01-15T10:35:00"
...
    InvoiceCreateRequest:
      type: object
      required:
        - paymentId
      properties:
        paymentId:
          type: integer
          format: int64
          example: 1
        description:
          type: string
          example: "Hóa đơn sửa điều hòa"
        itemsJson:
          type: string
          description: JSON string of invoice items
          example: '[{"description":"Thay gas R32","quantity":1,"unitPrice":300000.00}]'
    InvoiceResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        paymentId:
          type: integer
          format: int64
          example: 1
        invoiceNumber:
          type: string
          example: "INV-20240115-001"
        totalAmount:
          type: number
          format: double
          example: 500000.00
        taxAmount:
          type: number
          format: double
          example: 50000.00
        description:
          type: string
          example: "Hóa đơn sửa điều hòa"
        itemsJson:
          type: string
          example: '[{"description":"Thay gas R32","quantity":1,"unitPrice":300000.00}]'
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00"